requirements.txt
alembic==1.4.2
astroid==2.3.3
autopep8==1.5.1
click==7.1.1
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2-binary==2.8.5
pycodestyle==2.5.0
pylint==2.4.4
python-dateutil==2.8.1
python-editor==1.0.4
six==1.14.0
SQLAlchemy==1.3.16
typed-ast==1.4.1
Werkzeug==1.0.1
wrapt==1.11.2
